move tokenizer_utils to nlp
