+++ nemo/collections/nlp/modules/__init__.py
@@ -25,4 +25,6 @@ from nemo.collections.nlp.modules.common import (
     SequenceTokenClassifier,
     get_lm_model,
     get_pretrained_lm_models_list,
+    get_tokenizer,
+    get_tokenizer_list,
 )
