+++ nemo/collections/nlp/modules/common/__init__.py
@@ -15,15 +15,16 @@
 # limitations under the License.
 
 from nemo.collections.nlp.modules.common.bert_module import BertModule
-from nemo.collections.nlp.modules.common.common_utils import get_lm_model, get_pretrained_lm_models_list
 from nemo.collections.nlp.modules.common.huggingface import (
     AlbertEncoder,
     BertEncoder,
     DistilBertEncoder,
     RobertaEncoder,
 )
+from nemo.collections.nlp.modules.common.lm_utils import get_lm_model, get_pretrained_lm_models_list
 from nemo.collections.nlp.modules.common.megatron import MegatronBertEncoder
 from nemo.collections.nlp.modules.common.sequence_classifier import SequenceClassifier
 from nemo.collections.nlp.modules.common.sequence_regression import SequenceRegression
 from nemo.collections.nlp.modules.common.sequence_token_classifier import SequenceTokenClassifier
 from nemo.collections.nlp.modules.common.token_classifier import BertPretrainingTokenClassifier, TokenClassifier
+from nemo.collections.nlp.modules.common.tokenizer_utils import get_tokenizer, get_tokenizer_list
